Test: Share Vite plugins with metadata invalidation
