+++ test/unit/vitest.config.mjs
@@ -3,19 +3,17 @@ import { createRequire } from 'node:module';
 import path from 'node:path';
 import { fileURLToPath } from 'node:url';
 import { playwright } from '@vitest/browser-playwright';
-import react from '@vitejs/plugin-react-swc';
 import { globSync } from 'glob';
 import { defineConfig } from 'vitest/config';
+import { createVitePlugins } from './config/vite-plugins.mjs';
 import {
 	discoverTestFiles,
 	getVitestTestsByProject,
 } from './scripts/discover-test-files.mjs';
-import { compileInlineStyle } from '../../packages/wp-build/lib/compile-inline-style.mjs';
 
 const CONFIG_DIR = path.dirname( fileURLToPath( import.meta.url ) );
 const ROOT_DIR = path.resolve( CONFIG_DIR, '../..' );
 const nodeRequire = createRequire( import.meta.url );
-const emotionPlugin = nodeRequire.resolve( '@swc/plugin-emotion' );
 const gutenbergEnvSetupFile = path.join(
 	ROOT_DIR,
 	'test/unit/config/gutenberg-env.js'
@@ -37,54 +35,6 @@ const vitestTests = getVitestTestsByProject(
 const styleMockAlias = {
 	find: /^.*\.(?:css|scss)$/,
 	replacement: path.join( ROOT_DIR, 'test/unit/config/style-mock.vitest.js' ),
-};
-const WP_BUILD_CSS_MODULE_STYLE_FIXTURE_ID = 'virtual:wp-build-style-injection';
-const WP_BUILD_ORDINARY_STYLE_FIXTURE_ID =
-	'virtual:wp-build-ordinary-style-injection';
-const wpBuildStyleFixtureSources = new Map( [
-	[
-		WP_BUILD_CSS_MODULE_STYLE_FIXTURE_ID,
-		await compileInlineStyle( {
-			cssModules: true,
-			minify: false,
-		} )(
-			`@layer wp-build-test {
-				.fixture {
-					--wp-build-style-injection-test: true;
-					color: rgb(1, 2, 3);
-				}
-			}`,
-			ROOT_DIR,
-			path.join(
-				ROOT_DIR,
-				'test/unit/config/wp-build-style-fixture.module.css'
-			)
-		),
-	],
-	[
-		WP_BUILD_ORDINARY_STYLE_FIXTURE_ID,
-		await compileInlineStyle( { minify: false } )(
-			`.ordinary-fixture {
-				background-color: rgb(4, 5, 6);
-			}`,
-			ROOT_DIR,
-			path.join(
-				ROOT_DIR,
-				'test/unit/config/wp-build-ordinary-style-fixture.css'
-			)
-		),
-	],
-] );
-const wpBuildStyleFixturePlugin = {
-	name: 'wp-build-style-injection-fixture',
-	resolveId( id ) {
-		return wpBuildStyleFixtureSources.has( id ) ? `\0${ id }` : null;
-	},
-	load( id ) {
-		return id.startsWith( '\0' )
-			? wpBuildStyleFixtureSources.get( id.slice( 1 ) ) ?? null
-			: null;
-	},
 };
 const reporters = [ 'default' ];
 
@@ -130,20 +80,7 @@ export default defineConfig( {
 			runtime: 'automatic',
 		},
 	},
-	plugins: [
-		wpBuildStyleFixturePlugin,
-		react( {
-			plugins: [
-				[
-					emotionPlugin,
-					{
-						autoLabel: 'always',
-						labelFormat: '[local]',
-					},
-				],
-			],
-		} ),
-	],
+	plugins: await createVitePlugins( ROOT_DIR ),
 	resolve: {
 		alias: [
 			{
